+++ packages/eyes-sdk-core/lib/EyesBase.js
@@ -975,13 +975,13 @@ class EyesBase {
   /**
    * If a test is running, aborts it. Otherwise, does nothing.
    *
-   * @return {Promise<void>} A promise which resolves to the test results.
+   * @return {Promise<?TestResults>} A promise which resolves to the test results.
    */
   async abortIfNotClosed() {
     try {
       if (this._isDisabled) {
         this._logger.verbose('Eyes abortIfNotClosed ignored. (disabled)');
-        return;
+        return null;
       }
 
       this._isOpen = false;
@@ -991,17 +991,17 @@ class EyesBase {
 
       if (!this._runningSession) {
         this._logger.verbose('Closed');
-        return;
+        return null;
       }
 
       this._logger.verbose('Aborting server session...');
-      try {
-        // When aborting we do not save the test.
-        await this._serverConnector.stopSession(this._runningSession, true, false);
-        this._logger.log('--- Test aborted.');
-      } catch (err) {
-        this._logger.log(`Failed to abort server session: ${err}`);
-      }
+      const testResults = await this._serverConnector.stopSession(this._runningSession, true, false);
+
+      this._logger.log('--- Test aborted.');
+      return testResults;
+    } catch (err) {
+      this._logger.log(`Failed to abort server session: ${err}`);
+      return null;
     } finally {
       this._runningSession = null;
       this._logger.getLogHandler().close();
@@ -1060,11 +1060,11 @@ class EyesBase {
   /**
    * Sets the host OS name - overrides the one in the agent string.
    *
-   * @param {string} hostOS The host OS running the AUT.
+   * @param {string} hostOSInfo The host OS running the AUT.
    */
   setHostOSInfo(hostOSInfo) {
     this._logger.log(`Host OS Info: ${hostOSInfo}`);
-    if (hostOS) {
+    if (hostOSInfo) {
       this._hostOSInfo = hostOSInfo.trim();
     } else {
       this._hostOSInfo = undefined;
@@ -1083,7 +1083,7 @@ class EyesBase {
   /**
    * Sets the host application - overrides the one in the agent string.
    *
-   * @param {string} hostApp The application running the AUT (e.g., Chrome).
+   * @param {string} hostAppInfo The application running the AUT (e.g., Chrome).
    */
   setHostAppInfo(hostAppInfo) {
     this._logger.log(`Host App Info: ${hostAppInfo}`);
@@ -1106,7 +1106,7 @@ class EyesBase {
   /**
    * Sets the host application - overrides the one in the agent string.
    *
-   * @param {string} hostApp The application running the AUT (e.g., Chrome).
+   * @param {string} deviceInfo The application running the AUT (e.g., Chrome).
    */
   setDeviceInfo(deviceInfo) {
     this._logger.log(`Device Info: ${deviceInfo}`);