+++ packages/eyes-images-browser/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.4.0](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.3.1...@applitools/eyes-images-browser@4.4.0) (2018-12-04)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
+## [4.3.1](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.3.0...@applitools/eyes-images-browser@4.3.1) (2018-11-29)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
+# [4.3.0](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.2.0...@applitools/eyes-images-browser@4.3.0) (2018-11-29)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
 # [4.2.0](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.1.0...@applitools/eyes-images-browser@4.2.0) (2018-11-26)
 
 **Note:** Version bump only for package @applitools/eyes-images-browser