+++ packages/eyes-images/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.5.0](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images@4.4.0...@applitools/eyes-images@4.5.0) (2018-12-04)
+
+**Note:** Version bump only for package @applitools/eyes-images
+
+
+
+
+
+# [4.4.0](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images@4.3.0...@applitools/eyes-images@4.4.0) (2018-11-29)
+
+
+### Features
+
+* **eyes-images:** add `Target.imageProvider` method ([79d5ad9](https://github.com/applitools/eyes.sdk.javascript1/commit/79d5ad9))
+
+
+
+
+
+# [4.3.0](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images@4.2.0...@applitools/eyes-images@4.3.0) (2018-11-29)
+
+
+### Features
+
+* **eyes-images:** add withDom(domString) to checkSettings ([d6849d5](https://github.com/applitools/eyes.sdk.javascript1/commit/d6849d5))
+
+
+
+
+
 # [4.2.0](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images@4.1.0...@applitools/eyes-images@4.2.0) (2018-11-26)
 
 **Note:** Version bump only for package @applitools/eyes-images